Start refactoring DebugAdapter/DebugSession
