+++ src/adapter/IExecutionEngine.cs
@@ -1,10 +1,7 @@
 ﻿using Microsoft.VisualStudio.Shared.VSCodeDebugProtocol.Messages;
 using Neo.VM;
-using NeoDebug.Models;
 using NeoDebug.VariableContainers;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace NeoDebug
 {
