+++ src/adapter/InteropServices/InteropService.Blockchain.cs
@@ -1,10 +1,7 @@
 ﻿using Neo.VM;
 using NeoFx;
 using NeoFx.Models;
-using NeoFx.Storage;
 using System;
-using System.Collections.Generic;
-using System.Text;
