+++ src/adapter/InteropServices/InteropService.Runtime.cs
@@ -1,6 +1,5 @@
 ﻿using Microsoft.VisualStudio.Shared.VSCodeDebugProtocol.Messages;
 using Neo.VM;
-using NeoDebug;
 using System;
 using System.Collections.Generic;
 using System.Linq;
