+++ src/adapter/ModelAdapters/TransactionAdapter.cs
@@ -4,7 +4,6 @@ using NeoDebug.VariableContainers;
 using NeoFx;
 using NeoFx.Models;
 using NeoFx.Storage;
-using System;
 using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
