+++ src/adapter/Models/SequencePoint.cs
@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace NeoDebug.Models
+﻿namespace NeoDebug.Models
 {
     public class SequencePoint
     {
