+++ src/adapter/VariableContainers/ByteArrayContainer.cs
@@ -2,8 +2,6 @@
 using Neo.VM.Types;
 using System;
 using System.Collections.Generic;
-using System.Numerics;
-using System.Text;
 
 namespace NeoDebug.VariableContainers
 {
