+++ src/adapter/VariableContainers/IVariableContainer.cs
@@ -1,7 +1,5 @@
 ﻿using Microsoft.VisualStudio.Shared.VSCodeDebugProtocol.Messages;
-using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace NeoDebug.VariableContainers
 {
