+++ src/adapter/VariableContainers/IVariableContainerSession.cs
@@ -1,9 +1,4 @@
-﻿using Microsoft.VisualStudio.Shared.VSCodeDebugProtocol.Messages;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace NeoDebug.VariableContainers
+﻿namespace NeoDebug.VariableContainers
 {
     public interface IVariableContainerSession
     {
