+++ src/adapter/VariableContainers/IVariableProvider.cs
@@ -1,7 +1,4 @@
 ﻿using Microsoft.VisualStudio.Shared.VSCodeDebugProtocol.Messages;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
 namespace NeoDebug.VariableContainers
 {
